file restructure + add navigation
