+++ GeoNewsFinder4/package.json
@@ -11,15 +11,20 @@
   "dependencies": {
     "@expo/webpack-config": "^19.0.0",
     "@react-google-maps/api": "^2.19.2",
+    "@react-navigation/native": "^6.1.9",
+    "@react-navigation/stack": "^6.3.20",
     "expo": "~49.0.15",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.72.6",
     "react-native-maps": "1.7.1",
+    "react-native-safe-area-context": "4.6.3",
+    "react-native-screens": "~3.22.0",
     "react-native-web": "~0.19.6",
     "react-native-web-maps": "^0.3.0",
-    "react-native-webview": "^13.6.2"
+    "react-native-webview": "^13.6.2",
+    "react-native-gesture-handler": "~2.12.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
